Hardware querying fixes
Hardware IDs start out as 0xffff now, since it's unlikely there will
ever be 65,536 devices hooked up at the same time. It's a bad idea to
use zero (or anything that might correspond to a real device) because of
the following scenario:

1) Floppy drive is device zero (or whatever).
2) Clock is not found, so its device ID is never changed from zero (or
whatever).
3) We do something with the clock, but it's really doing something
random with the floppy, perhaps destroying data.