Even More Calculator Additions and fixes
The new calculator will not allow you to make mathematical errors like
6-+6 or 6//6 without creating an error.
Also, you can now set the base to a number by adding a tag at the end
of it (10h) for 10 in hex.
The tags are i for b"i"nary, o for "o"ctal, t for "t"en, and h for "h"ex
I couldn't use b or d as tags because it would cause confusion with
hex's b and d.
A new command "setDefaultBase" was also added. This will allow you to
change the default base between the number 2 and 16. Always assigned in
decimal
So if I type, base 2, then type 10 on the next line, 2 would be the
value since 10 in binary is 2.
mod ( % ) operator was added.
You can "double fib still" ( !!7 = !13 = 233)

Still no order of operations.  May take a while for that

Example Used:
Base is by default 10.
!!7%fh^10i
233%15^2
8^2
64